delegated-ink/exception-thrown-bad-color.tentative.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/resources/testdriver-vendor.js"></script>

<canvas id="canvas"></canvas>
<script>
  let presenter = navigator.ink.requestPresenter('delegated-ink-trail', canvas);
  let style = { color: "bad-color", diameter: 6 };

  canvas.addEventListener("pointermove", evt => {
      presenter.then( function(p) {
        test(() => {
          assert_throws_js(TypeError, function() {
            p.updateInkTrailStartPoint(evt, style);
          }, "TypeError is expected with a bad color");
        }, "Expected a TypeError to be thrown due to bad color.");
      });
  });

  new test_driver.Actions().pointerMove(10,10).send();
</script>
